requirements.txt
# Core LLM and Agent Framework
langchain==0.1.0
langchain-community==0.0.13
langchain-huggingface==0.0.1
huggingface_hub==0.20.2

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# API and HTTP
requests==2.31.0
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
rich==13.7.0
colorama==0.4.6

# Date/Time handling
python-dateutil==2.8.2

# JSON Schema validation
jsonschema==4.20.0

